serialization: Better serialization
Now components doesn't have to inherit from SerializableComponent but
need to have only specified attributes set to class/fields.
SerializationManager handle all the stuff. SerializationManager
must be created for each scene we want to serialize data.

+++ Assets/Editor/InitOnLoad.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 using UnityEditor;
-using System.Collections;
+using System.Collections.Generic;
 
 //! Handle the editor-on-load stuff and assets modifications.
 [InitializeOnLoad]
@@ -16,8 +16,16 @@ public class InitOnLoad : UnityEditor.AssetModificationProcessor
 	//! Hande pre-save action.
 	public static string[] OnWillSaveAssets(string[] Paths)
 	{
-		// Try to serialize objects
-		SendSerialize();
+		// Try to serialize objects if there is serialization manager in scene
+		// If there is serialization manager in scene.
+		var serialization_manager = (SerializationManager)GameObject.FindObjectOfType(typeof(SerializationManager));
+		if (serialization_manager)
+		{
+			serialization_manager.Serialize();
+			// Set as dirty
+			EditorUtility.SetDirty(serialization_manager);
+		}
+		
 
 		return Paths;
 	}
@@ -25,43 +33,23 @@ public class InitOnLoad : UnityEditor.AssetModificationProcessor
 	//! Handle the editor -> player change.
 	public static void Change()
 	{
-		// Editor -> Player
-		if (EditorApplication.isPlayingOrWillChangePlaymode && !EditorApplication.isPlaying)
+		var serialization_manager = (SerializationManager)GameObject.FindObjectOfType(typeof(SerializationManager));
+		if (serialization_manager)
 		{
-			Debug.Log("Change in EDIT");
-			SendSerialize();
-			// Touch file to signal we're going to player
-			System.IO.File.Create(Application.temporaryCachePath + "/" + TEMP_FILE_NAME);
-		}
-		// Player -> Editor
-		else if(!EditorApplication.isPlayingOrWillChangePlaymode && EditorApplication.isPlaying)
-		{
-			// Remove file to signal we're going to editor
-			System.IO.File.Delete(Application.temporaryCachePath + "/" + TEMP_FILE_NAME);
-		}
-	}
-
-	//! Try to serialize all game objects.
-	static void SendSerialize()
-	{
-		Object[] gos = GameObject.FindObjectsOfType(typeof(GameObject));
-		foreach (GameObject go in gos)
-		{
-			var serialize_component = go.GetComponent<SerializableComponent>();
-			if (serialize_component != null)
-				serialize_component.Serialize();
-		}
-	}
-
-	//! Try to serialize all game objects.
-	static void SendDeSerialize()
-	{
-		Object[] gos = GameObject.FindObjectsOfType(typeof(GameObject));
-		foreach (GameObject go in gos)
-		{
-			var serialize_component = go.GetComponent<SerializableComponent>();
-			if (serialize_component != null)
-				serialize_component.Deserialize();
+			// Editor -> Player
+			if (EditorApplication.isPlayingOrWillChangePlaymode && !EditorApplication.isPlaying)
+			{
+				Debug.Log("Change in EDIT");
+				serialization_manager.Serialize();
+				// Touch file to signal we're going to player
+				System.IO.File.Create(Application.temporaryCachePath + "/" + TEMP_FILE_NAME);
+			}
+			// Player -> Editor
+			else if (!EditorApplication.isPlayingOrWillChangePlaymode && EditorApplication.isPlaying)
+			{
+				// Remove file to signal we're going to editor
+				System.IO.File.Delete(Application.temporaryCachePath + "/" + TEMP_FILE_NAME);
+			}
 		}
 	}
 #endregion
@@ -70,22 +58,34 @@ public class InitOnLoad : UnityEditor.AssetModificationProcessor
 	static InitOnLoad()
 	{
 		Debug.Log("InitOnLoad loaded...");
+		
 		EditorApplication.playmodeStateChanged += Change;
-		// InitOnLoad() is ran not only on editor startup but also if we're in
-		// editor and some code needs to be reloaded. Therefor, all
-		// components in editor are default-serialized by unity, without
-		// executing our serialization code. We must the explicitly call
-		// deserializatin code.
-		try
+
+		// If there is serialization manager in scene.
+		var serialization_manager = (SerializationManager)GameObject.FindObjectOfType(typeof(SerializationManager));
+		if (serialization_manager)
 		{
-			// Only if not in player, otherwise double deserialization occurs
-			// (one here, another one in Awake() of serialization component)
-			if(!System.IO.File.Exists(Application.temporaryCachePath + "/" + TEMP_FILE_NAME))
-				SendDeSerialize();
-		}
-		catch(System.Exception e)
-		{
-			Debug.LogWarning("Exception while deserializing on init:" + e.ToString());
+			// InitOnLoad() is ran not only on editor startup but also if we're in
+			// editor and some code needs to be reloaded. Therefor, all
+			// components in editor are default-serialized by unity, without
+			// executing our serialization code. We must the explicitly call
+			// deserialization code.
+			try
+			{
+				// Only if not in player, otherwise double deserialization occurs
+				// (one here, another one in Awake() of serialization component)
+				// Also reload the type infos here, because assemblies could be
+				// changed here
+				if (!System.IO.File.Exists(Application.temporaryCachePath + "/" + TEMP_FILE_NAME))
+				{
+					serialization_manager.Reload();
+					serialization_manager.Deserialize();
+				}
+			}
+			catch (System.Exception e)
+			{
+				Debug.LogWarning("Exception while deserializing on init:" + e.ToString());
+			}
 		}
 	}
 #endregion
